add telnet, update scrapli pin, break things up a bit
