+++ .gitignore
@@ -1,4 +1,5 @@
 .vscode
+.idea
 outputs*
 dist
 cmdo
