+++ go.mod
@@ -4,7 +4,7 @@ go 1.16
 
 require (
 	github.com/fatih/color v1.12.0
-	github.com/scrapli/scrapligo v0.0.0-20210704164516-6c3b4e74cfad
+	github.com/scrapli/scrapligo v0.0.0-20210820234601-7f17be656249
 	github.com/sirupsen/logrus v1.8.1
 	github.com/srl-labs/srlinux-scrapli v0.0.0-20210601201111-9eed8d440381
 	github.com/urfave/cli/v2 v2.3.0
